imporve compoent to refactorist
