+++ features/kasir/components/form/PaymentSummary/index.ts
@@ -1,0 +1,7 @@
+// Export barrel for PaymentSummary components
+export { PaymentBreakdownSection } from './PaymentBreakdownSection'
+export { DiscountSection } from './DiscountSection'
+export { PaymentMethodSection } from './PaymentMethodSection'
+export { OrderSummarySection } from './OrderSummarySection'
+export { RentalDurationSection } from './RentalDurationSection'
+export { NotesSection } from './NotesSection'
